requirements.txt
langchain==0.1.5
langchain_community==0.0.17
langchain_core==0.1.18
langchain_openai==0.0.5
python-dotenv==1.0.1
streamlit==1.31.0
beautifulsoup4==4.12.2
faiss-cpu==1.7.4


# this file created by command : pipreqs /Users/srm/pyt/web-chat   